+++ include/opcontrol.hpp
@@ -23,13 +23,13 @@ using namespace std;
 #define SCALING 1.0           // Constant that all chassis speeds are multiplied by, 
                               // used to scale all chassis movement. 
 
-void processInput(int * arrInputs);
+void processInput(int8_t * arrInputs);
 
 void startRecordThread();
 
 void recordLoop(void * param);
 
-void readController(int * instInputs);
+void readController(int8_t * instInputs);
 
 // Helper for calculating averages
 int average(int x, int y);
